Move RenderTypes
