+++ src/main/java/moriyashiine/aylyth/client/render/entity/living/layer/WreathedHindGlowLayerRenderer.java
@@ -1,6 +1,6 @@
 package moriyashiine.aylyth.client.render.entity.living.layer;
 
-import moriyashiine.aylyth.client.RenderTypes;
+import moriyashiine.aylyth.client.render.RenderTypes;
 import moriyashiine.aylyth.common.entity.mob.WreathedHindEntity;
 import moriyashiine.aylyth.common.util.AylythUtil;
 import net.fabricmc.api.EnvType;
